create a new class: PersistenceManager
